ref: move lr_finder (#3434)
* ref: move lr_finder

* ref: move lr_finder

* ref: move lr_finder

* ref: move lr_finder

* ref: move lr_finder

* ref: move lr_finder

* ref: move lr_finder
