+++ pytorch_lightning/tuner/tuning.py
@@ -13,6 +13,10 @@
 # limitations under the License.
 from pytorch_lightning.tuner.batch_size_scaling import scale_batch_size
 from pytorch_lightning.tuner.auto_gpu_select import pick_multiple_gpus
+from pytorch_lightning.tuner.lr_finder import _run_lr_finder_internally, lr_find
+from pytorch_lightning.core.lightning import LightningModule
+from typing import Optional, List, Union
+from torch.utils.data import DataLoader
 
 
 class Tuner:
@@ -31,6 +35,32 @@ class Tuner:
         return scale_batch_size(
             self.trainer, model, mode, steps_per_trial, init_val, max_trials, batch_arg_name, **fit_kwargs
         )
+
+    def lr_find(
+            self,
+            model: LightningModule,
+            train_dataloader: Optional[DataLoader] = None,
+            val_dataloaders: Optional[Union[DataLoader, List[DataLoader]]] = None,
+            min_lr: float = 1e-8,
+            max_lr: float = 1,
+            num_training: int = 100,
+            mode: str = 'exponential',
+            early_stop_threshold: float = 4.0,
+    ):
+        return lr_find(
+            self.trainer,
+            model,
+            train_dataloader,
+            val_dataloaders,
+            min_lr,
+            max_lr,
+            num_training,
+            mode,
+            early_stop_threshold
+        )
+
+    def internal_find_lr(self, trainer, model: LightningModule):
+        return _run_lr_finder_internally(trainer, model)
 
     def pick_multiple_gpus(self, num_gpus: int):
         return pick_multiple_gpus(num_gpus)
